Move clusterHealthResponse into own file called cluster_health_response.go
