+++ collector/cluster_health.go
@@ -36,25 +36,6 @@ type ClusterHealth struct {
 	UnassignedShards        *prometheus.Desc
 }
 
-// Generated with https://mholt.github.io/json-to-go/
-type clusterHealthResponse struct {
-	ClusterName                 string  `json:"cluster_name"`
-	Status                      string  `json:"status"`
-	TimedOut                    bool    `json:"timed_out"`
-	NumberOfNodes               int     `json:"number_of_nodes"`
-	NumberOfDataNodes           int     `json:"number_of_data_nodes"`
-	ActivePrimaryShards         int     `json:"active_primary_shards"`
-	ActiveShards                int     `json:"active_shards"`
-	RelocatingShards            int     `json:"relocating_shards"`
-	InitializingShards          int     `json:"initializing_shards"`
-	UnassignedShards            int     `json:"unassigned_shards"`
-	DelayedUnassignedShards     int     `json:"delayed_unassigned_shards"`
-	NumberOfPendingTasks        int     `json:"number_of_pending_tasks"`
-	NumberOfInFlightFetch       int     `json:"number_of_in_flight_fetch"`
-	TaskMaxWaitingInQueueMillis int     `json:"task_max_waiting_in_queue_millis"`
-	ActiveShardsPercentAsNumber float64 `json:"active_shards_percent_as_number"`
-}
-
 func NewClusterHealth(logger log.Logger, client *http.Client, url *url.URL) *ClusterHealth {
 	subsystem := "cluster_health"
